Refactor: move files 파일 구조 정리를 위해 data 디렉터리의 data, repository 파일을 domain/data의 하위로 이동. 추후 domain의 naming 변경 예정
